requirements.txt
arrow==1.2.1
autopep8==1.6.0
azure-functions==1.8.0
backports.entry-points-selectable==1.1.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
distlib==0.3.3
docopt==0.6.2
docutils==0.19
filelock==3.3.2
idna==3.3
mypy-extensions==0.4.3
numpy==1.24.1
panda==0.3.1
pandas==1.5.2
pathspec==0.9.0
pep8==1.7.1
pipreqs==0.4.11
pipreqsnb==0.2.4
platformdirs==2.4.0
protobuf==4.21.9
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
stravalib==0.10.4
tabulate==0.8.9
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
units==0.7
urllib3==1.26.7
virtualenv==20.10.0
yarg==0.1.9
